refactor: consolidate core tag and asset modules
Merge tag/parser/tag_expand/autodiscover into core/tags.py and
render_assets into core/assets.py; break the assets<->finder import
cycle; de-dup the asset manifest builder. Behavior unchanged.
